requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
dateparser==1.1.4
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
mysql-connector-python==8.0.31
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
plotly==5.11.0
protobuf==3.20.1
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.1
six==1.16.0
tenacity==8.1.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.13
websockets==10.4
yarl==1.8.2
